Removed warnings on aster.

+++ trunk/core/src/clbinfo.c
@@ -94,13 +94,13 @@ int main(int argc, char *argv[])
 
   fp = clfopen(outfile,"a+");
   fprintf(fp, "%d %d %f %f %f %f %f\n",
-      dm->size1, maxtrials, (double) matched / (double) maxtrials,
+      (int) dm->size1, (int) maxtrials, (double) matched / (double) maxtrials,
       finish - start, (finish - start) / (2.0 * maxtrials), log((finish-start)/(2.0*maxtrials))/log(2.0), log(finish-start)/log(2.0));
 
   printf("matched: %d\n", matched);
   printf("maxtrials: %d\n", maxtrials);
   printf("%.2f%% matching tree pairs for %dx%d matrix\n",
-      ( 100.0 * (double) matched / (double) maxtrials), dm->size1, dm->size1);
+      ( 100.0 * (double) matched / (double) maxtrials), (int) dm->size1, (int) dm->size1);
 
   clfclose(fp);
   gsl_matrix_free(dm);
